Refactor: Build the examples import map in a pure helper
